+++ examples/features/codes.rs
@@ -5,9 +5,7 @@ use termal::{
     error::Result,
     formatc, printcln,
     raw::{
-        TermSize, Terminal, disable_raw_mode, enable_raw_mode,
-        events::{Event, Key, KeyCode, Modifiers},
-        term_size,
+        disable_raw_mode, enable_raw_mode, events::{Event, Key, KeyCode, Modifiers, StateChange}, term_size, TermSize, Terminal
     },
     reset_terminal,
 };
@@ -1288,5 +1286,37 @@ pub fn show_limit_print_to_scroll_region() -> Result<()> {
     print!("{}", codes::DONT_LIMIT_PRINT_TO_SCROLL_REGION);
     print!("{}", codes::RESET_SCROLL_REGION);
 
+    Ok(())
+}
+
+pub fn show_enable_bracketed_paste_mode() -> Result<()> {
+    enable_raw_mode()?;
+    
+    let mut term = Terminal::stdio();
+    
+    print!("{}", codes::CLEAR);
+    print!("{}", codes::ENABLE_BRACKETED_PASTE_MODE);
+    term.println("With bracketed paste mode:")?;
+    
+    loop {
+        let e = term.read()?;
+        term.println(format!("{e:?}"))?;
+        if matches!(e, Event::StateChange(StateChange::BracketedPasteEnd)) {
+            break;
+        }
+    }
+    
+    term.flushed(format!("{}", codes::DISABLE_BRACKETED_PASTE_MODE))?;
+    term.println("Without brakceted paste mode:")?;
+    
+    let e = term.read()?;
+    term.println(format!("{e:?}"))?;
+    while term.has_input() {
+        let e = term.read()?;
+        term.println(format!("{e:?}"))?;
+    }
+    
+    disable_raw_mode()?;
+    
     Ok(())
 }