+++ termal_core/src/codes.rs
@@ -4829,7 +4829,112 @@ pub const LIMIT_PRINT_TO_SCROLL_REGION: &str = disable!(19);
 
 /// Enables bracketed paste mode. In this mode, pasted text is treated
 /// verbatim.
+/// 
+/// Equivalent to `CSI ? 2 0 0 4 h`.
+/// 
+/// This mode may be disabled with [`DISABLE_BRACKETED_PASTE_MODE`].
+/// 
+/// If this mode is enabled, pasted text will be prepended with `CSI 2 0 0 ~`
+/// ([`BRACKETED_PASTE_START`]) and appended with `CSI 2 0 1 ~`
+/// ([`BRACKETED_PASTE_END`]).
+/// 
+/// Termal can parse these events as [`crate::raw::events::Event::StateChange`]
+/// with either [`crate::raw::events::StateChange::BracketedPasteStart`] and
+/// [`crate::raw::events::StateChange::BracketedPasteEnd`]. Terminal will treat
+/// all text in between these events verbatim.
+/// 
+/// # Example
+/// ```no_run
+/// use termal_core::{
+///     codes,
+///     raw::{
+///         enable_raw_mode, disable_raw_mode, Terminal,
+///         events::{Event, StateChange},
+///     }
+/// };
+/// 
+/// enable_raw_mode()?;
+/// 
+/// let mut term = Terminal::stdio();
+/// 
+/// print!("{}", codes::CLEAR);
+/// print!("{}", codes::ENABLE_BRACKETED_PASTE_MODE);
+/// term.println("With bracketed paste mode:")?;
+/// 
+/// loop {
+///     let e = term.read()?;
+///     term.println(format!("{e:?}"))?;
+///     if matches!(e, Event::StateChange(StateChange::BracketedPasteEnd)) {
+///         break;
+///     }
+/// }
+/// 
+/// term.flushed(format!("{}", codes::DISABLE_BRACKETED_PASTE_MODE))?;
+/// term.println("Without brakceted paste mode:")?;
+/// 
+/// let e = term.read()?;
+/// term.println(format!("{e:?}"))?;
+/// while term.has_input() {
+///     let e = term.read()?;
+///     term.println(format!("{e:?}"))?;
+/// }
+/// 
+/// disable_raw_mode()?;
+/// 
+/// Ok(())
+/// ```
+///
+/// ## Result in terminal
+/// ![](https://raw.githubusercontent.com/BonnyAD9/termal/refs/heads/master/assets/codes/enable_bracketed_paste_mode.png)
 pub const ENABLE_BRACKETED_PASTE_MODE: &str = enable!(2004);
+/// Disables bracketed paste mode that was enabled with
+/// [`ENABLE_BRACKETED_PASTE_MODE`].
+/// 
+/// Equivalent to `CSI ? 2 0 0 4 l`.
+/// 
+/// # Example
+/// ```no_run
+/// use termal_core::{
+///     codes,
+///     raw::{
+///         enable_raw_mode, disable_raw_mode, Terminal,
+///         events::{Event, StateChange},
+///     }
+/// };
+/// 
+/// enable_raw_mode()?;
+/// 
+/// let mut term = Terminal::stdio();
+/// 
+/// print!("{}", codes::CLEAR);
+/// print!("{}", codes::ENABLE_BRACKETED_PASTE_MODE);
+/// term.println("With bracketed paste mode:")?;
+/// 
+/// loop {
+///     let e = term.read()?;
+///     term.println(format!("{e:?}"))?;
+///     if matches!(e, Event::StateChange(StateChange::BracketedPasteEnd)) {
+///         break;
+///     }
+/// }
+/// 
+/// term.flushed(format!("{}", codes::DISABLE_BRACKETED_PASTE_MODE))?;
+/// term.println("Without brakceted paste mode:")?;
+/// 
+/// let e = term.read()?;
+/// term.println(format!("{e:?}"))?;
+/// while term.has_input() {
+///     let e = term.read()?;
+///     term.println(format!("{e:?}"))?;
+/// }
+/// 
+/// disable_raw_mode()?;
+/// 
+/// Ok(())
+/// ```
+///
+/// ## Result in terminal
+/// ![](https://raw.githubusercontent.com/BonnyAD9/termal/refs/heads/master/assets/codes/enable_bracketed_paste_mode.png)
 pub const DISABLE_BRACKETED_PASTE_MODE: &str = disable!(2004);
 
 #[derive(Clone, Debug, Copy, Eq, PartialEq)]